+++ internal/utils/handler/dto/save_route.go
@@ -2,7 +2,7 @@ package dto
 
 //go:generate ffjson -noencoder $GOFILE
 type SaveRouteDto struct {
-	SourceID      int64   `json:"source"`
-	DestinationID int64   `json:"destination"`
-	Price         float32 `json:"price"`
+	SourceID      int64 `json:"source"`
+	DestinationID int64 `json:"destination"`
+	Price         int32 `json:"price"`
 }
